feat: Add AGUI processors for current turn normalization, persisted context handling, and plan version selection
- Implemented AgUiCurrentTurnProcessor to normalize user messages in AGUI.
- Created AgUiPersistedContextCaptureProcessor and AgUiPersistedContextProcessor for managing conversation context persistence.
- Added AgUiPlanVersionSelectorProcessor to select plan versions based on incoming parameters and headers.
- Introduced ConversationContextExtractor, Merger, and Renderer interfaces for flexible context handling.
- Developed PersistedConversationContextStore for managing conversation context persistence with fallback mechanisms.
- Implemented RealtimePersistedContextCaptureProcessor and RealtimePersistedContextProcessor for real-time context management.
- Created OpenAiRealtimeCallProcessor for handling OpenAI API calls with multipart request support.
- Added TelephonyIdentityMetadataProcessor to extract and promote caller identity metadata.
- Introduced TwilioInboundCallerGateProcessor to validate inbound callers against a configured allowlist.
- Added unit tests for OpenAiRealtimeCallProcessor and TelephonyIdentityMetadataProcessor to ensure functionality.
- Implemented PhoneNumberSupport utility for normalizing phone numbers.

+++ camel-agent-spring-ai/src/test/java/io/dscope/camel/agent/springai/MultiProviderSpringAiChatGatewayTest.java
@@ -152,7 +152,7 @@ class MultiProviderSpringAiChatGatewayTest {
             .build();
 
         Assertions.assertNull(options.getMaxTokens());
-        Assertions.assertEquals(900, options.getMaxCompletionTokens());
+        Assertions.assertEquals(900, options.getMaxCompletionTokens().intValue());
     }
 
     @Test
@@ -161,7 +161,7 @@ class MultiProviderSpringAiChatGatewayTest {
             .applyOpenAiMaxTokenOption(OpenAiChatOptions.builder(), "gpt-4o", 700)
             .build();
 
-        Assertions.assertEquals(700, options.getMaxTokens());
+        Assertions.assertEquals(700, options.getMaxTokens().intValue());
         Assertions.assertNull(options.getMaxCompletionTokens());
     }
